+++ api/helpers/keys/generateKey.js
@@ -1,0 +1,5 @@
+function generateSecret() {
+    return Math.floor((Math.random() * 999999) + 100000);
+}
+
+exports.generateSecret = generateSecret;
